Flask API/api/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
urllib3==1.26.9
Werkzeug==2.2.2
PyJWT==2.3.0
pyodbc==4.0.34
gunicorn==20.1.0
mysqlclient==2.1.0
PyMySQL==1.0.2
click==8.1.3
colorama==0.4.5
Jinja2==3.1.2
MarkupSafe==2.1.1
itsdangerous==2.1.2